+++ ClashGui/Clash/Models/Connections/ConnectionInfo.cs
@@ -4,10 +4,10 @@ namespace ClashGui.Clash.Models.Connections;
 
 public class ConnectionInfo
 {
-    public int DownloadTotal { get; set; }
+    public long DownloadTotal { get; set; }
 
 
-    public int UploadTotal { get; set; }
+    public long UploadTotal { get; set; }
 
 
     public List<Connection> Connections { get; set; }
